+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 * Fixes pydantic issue in `XMLiveChannel`
+* Adjusts selected HLS stream to (hopefully) fix `radio_time`
 
 0.2.3 (2021-08-15)
 ------------------